+ Delete user(s)

+++ app/Models/Forum/Delete.php
@@ -4,6 +4,7 @@ namespace ForkBB\Models\Forum;
 
 use ForkBB\Models\Action;
 use ForkBB\Models\Forum\Model as Forum;
+use ForkBB\Models\User\Model as User;
 use InvalidArgumentException;
 use RuntimeException;
 
@@ -20,14 +21,23 @@ class Delete extends Action
     public function delete(...$args)
     {
         if (empty($args)) {
-            throw new InvalidArgumentException('No arguments, expected forum(s)');
+            throw new InvalidArgumentException('No arguments, expected User(s) or Forum(s)');
         }
 
-        $forums = [];
-        $all    = [];
+        $users   = [];
+        $forums  = [];
+        $all     = [];
+        $isUser  = 0;
+        $isForum = 0;
 
         foreach ($args as $arg) {
-            if ($arg instanceof Forum) {
+            if ($arg instanceof User) {
+                if ($arg->isGuest) {
+                    throw new RuntimeException('Guest can not be deleted');
+                }
+                $users[] = $arg->id;
+                $isUser  = 1;
+            } elseif ($arg instanceof Forum) {
                 if (! $this->c->forums->get($arg->id) instanceof Forum) {
                     throw new RuntimeException('Forum unavailable');
                 }
@@ -36,9 +46,14 @@ class Delete extends Action
                 foreach (\array_keys($arg->descendants) as $id) { //???? а если не админ?
                     $all[$id] = true;
                 }
+                $isForum = 1;
             } else {
-                throw new InvalidArgumentException('Expected forum(s)');
+                throw new InvalidArgumentException('Expected User(s) or Forum(s)');
             }
+        }
+
+        if ($isUser + $isForum > 1) {
+            throw new InvalidArgumentException('Expected only User(s) or Forum(s)');
         }
 
         if (\array_diff_key($all, $forums)) {
@@ -47,21 +62,33 @@ class Delete extends Action
 
         $this->c->topics->delete(...$args);
 
-        //???? подписки, опросы, предупреждения, метки посещения тем
+        //???? подписки, опросы, предупреждения
 
-        foreach ($forums as $forum) {
-            $this->c->groups->Perm->reset($forum);
+        if ($users) {
+            $vars = [
+                ':users' => $users,
+            ];
+            $sql = 'DELETE FROM ::mark_of_forum
+                    WHERE uid IN (?ai:users)';
+            $this->c->DB->exec($sql, $vars);
+
+            //???? удаление модераторов из разделов
         }
+        if ($forums) {
+            foreach ($forums as $forum) {
+                $this->c->groups->Perm->reset($forum);
+            }
 
-        $vars = [
-            ':forums' => \array_keys($forums),
-        ];
-        $sql = 'DELETE FROM ::mark_of_forum
-                WHERE fid IN (?ai:forums)';
-        $this->c->DB->exec($sql, $vars);
+            $vars = [
+                ':forums' => \array_keys($forums),
+            ];
+            $sql = 'DELETE FROM ::mark_of_forum
+                    WHERE fid IN (?ai:forums)';
+            $this->c->DB->exec($sql, $vars);
 
-        $sql = 'DELETE FROM ::forums
-                WHERE id IN (?ai:forums)';
-        $this->c->DB->exec($sql, $vars);
+            $sql = 'DELETE FROM ::forums
+                    WHERE id IN (?ai:forums)';
+            $this->c->DB->exec($sql, $vars);
+        }
     }
 }
